test: start adding integration tests for optimization on Guppy generated hugr programs (#1384)
I've started adding some basic tests that validate that `tket`
optimizations transform guppy programs as expected. This is part of
https://github.com/Quantinuum/guppylang/issues/1383

Note that some key use cases are not supported yet including

1. Some sequences of parameterized gates are not simplified as expected.
Hopefully once https://github.com/Quantinuum/hugr/pull/2739 is released
and used in `tket` this will be fixed.
2. Programs with arrays don't always simplify as expected (see
https://github.com/Quantinuum/tket2/issues/1307)
3. Guppy functions which call other functions will only be optimized at
the top level currently.

Once these limitations are removed I think it would be good to add test
cases for the above as well.