fix: Support comptime expressions in generic argument applications (#1409)
When calling a comptime expression in a subscript (such as a generic
argument application), the AST visitor happens to compile the `comptime`
expression early, i.e. the node given to the parsing module is already a
`ComptimeExpr` rather than an `ast.Call`. This made the
`is_comptime_expr` check fail, since it did not expect this conversion
to already have happened.

Closes #1087